Auth_Microservice_Completed
A token was issued to the user who logged in with jwt, and integration with the refresh token was achieved. Security was enhanced using Microsoft Identity. The registered user information and role were successfully added to the database. The project was launched via Docker and is being managed with Dbeaver.